Tela novo usuário finalizada e alertas
