+++ src/App.js
@@ -1,10 +1,14 @@
-import { VendaStorage } from './contexts/VendaContext';
+import { ToastContainer } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
 import { AuthProvider } from './contexts/Auth/AuthProvider';
-import Rotas from './pages/Rotas';
+import { VendaStorage } from './contexts/VendaContext';
 import './index.css';
+import Rotas from './pages/Rotas';
 
 const App = () => {
 	return (
+		<>
+		<ToastContainer />
 		<div className="divlanding">
 			<AuthProvider>
 				<VendaStorage>
@@ -12,6 +16,7 @@ const App = () => {
 				</VendaStorage>
 			</AuthProvider>
 		</div>
+		</>
 	);
 };
