+++ src/pages/Usuarios/UsuarioLista/UsuarioLista.js
@@ -15,7 +15,7 @@ useEffect(() => {
   fetch(`https://localhost:7171/api/Usuario?nome=sem%20nome&login=sem%20login&tamanho=0&pagina=1`, { 
     method: 'GET', 
     headers: new Headers({
-    'Authorization': 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1laWQiOiIxIiwibmFtZSI6IkFkbWluIiwiZW1haWwiOiJhZG1pbkBnbWFpbC5jb20iLCJpc19hZG1pbiI6IlRydWUiLCJuYmYiOjE2NTk3MTc4NDEsImV4cCI6MTY1OTcyNTA0MSwiaWF0IjoxNjU5NzE3ODQxfQ.Gcsec8YSdg8_xQeJ0apMlCVtV3MJCMirCJ4AsdSZzoo', 
+    'Authorization': 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1laWQiOiIxIiwibmFtZSI6IkFkbWluIiwiZW1haWwiOiJhZG1pbkBnbWFpbC5jb20iLCJpc19hZG1pbiI6IlRydWUiLCJuYmYiOjE2NTk3Mzc5MDAsImV4cCI6MTY1OTc0NTEwMCwiaWF0IjoxNjU5NzM3OTAwfQ.i8UzBLreSsYxD_ms4aLosQ4oMQjTsku5TjZCJlArAn0', 
     'Content-Type': 'application/json',
     }), 
   }).then(response => {
