+++ src/services/UsuariosService.js
@@ -3,7 +3,7 @@ export const getUsuarios = async (nome = 'sem nome', login = 'sem login', tamanh
     const reponse = await fetch(`https://localhost:7171/api/Usuario?nome=${nome}&login=${login}&tamanho=${tamanho}&pagina=${pagina}`, { 
                         method: 'GET', 
                         headers: new Headers({
-                        'Authorization': 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1laWQiOiIxIiwibmFtZSI6IkFkbWluIiwiZW1haWwiOiJhZG1pbkBnbWFpbC5jb20iLCJpc19hZG1pbiI6IlRydWUiLCJuYmYiOjE2NTk3MTc4NDEsImV4cCI6MTY1OTcyNTA0MSwiaWF0IjoxNjU5NzE3ODQxfQ.Gcsec8YSdg8_xQeJ0apMlCVtV3MJCMirCJ4AsdSZzoo', 
+                        'Authorization': `bearer ${localStorage.getItem('token')}`,
                         'Content-Type': 'application/json'
                         }), 
                     }).then(response => {
